==> 6th Place - Click Click Boom/zindi_sentinel2_codes/step5.2_extract_lat_long_test.ipynb <==
!pip install scikit-learn==1.0
!pip install xgboost==1.4.2
!pip install catboost==0.26.1
!pip install pandas==1.3.3
!pip install radiant-mlhub==0.3.0
!pip install rasterio==1.2.8
!pip install numpy==1.21.2
!pip install pathlib==1.0.1
!pip install tqdm==4.62.3
!pip install joblib==1.0.1
!pip install matplotlib==3.4.3
!pip install Pillow==8.3.2
!pip install torch==1.9.1
!pip install plotly==5.3.1


import warnings
warnings.filterwarnings('ignore')
# warnings.filterwarnings('RuntimeWarning')

from radiant_mlhub import Collection
import tarfile
import os
from pathlib import Path
import json
from tqdm import tqdm
from joblib import Parallel,delayed
import datetime
import rasterio
import numpy as np
import pandas as pd

import gc

gc.collect()

competition_train_df = pd.read_csv('test_data_sentinel2.csv')

def get_date_format(month,day):
    '''
        Structures the dates in a particular format
    '''
    if (str(month)=='nan') or (str(day)=='nan'):
        return 'nan'
    else:
        if month>=10:
            if day>=10:
                return f'month_{str(int(month))}_day_{str(int(day))}'
            else:
                return f'month_{str(int(month))}_day_0{str(int(day))}'
        else:
            if day>=10:
                return f'month_0{str(int(month))}_day_{str(int(day))}'
            else:
                return f'month_0{str(int(month))}_day_0{str(int(day))}'
    

competition_train_df['month'] = pd.to_datetime(competition_train_df['datetime']).dt.month.values 
competition_train_df['day']   = pd.to_datetime(competition_train_df['datetime']).dt.day.values

competition_train_df['dates'] = competition_train_df.apply(lambda z: get_date_format(z['month'],z['day']),axis=1)

unique_dates = competition_train_df['dates'].unique()
unique_dates = np.array([z for z in unique_dates if 'nan' not in z])

print(f'Length of unique dates {len(unique_dates)}')

date_dict = dict(zip(competition_train_df['datetime'].dropna().unique(),unique_dates))
date_dict = dict(sorted(date_dict.items(), key=lambda item: item[1]))
date_order_to_consider = np.array(list(date_dict.values()))
tile_ids_train = competition_train_df['tile_id'].unique()

def get_bands(tile_date_times,tile_df,band,date_dict):
    '''
        Getting band dictionary with dates
    '''
    X_tile    = np.zeros((256 * 256, 76))
    X_tile[:] = np.nan
    
    for date_time in tile_date_times[:1]:
        
        
        source   = rasterio.open(tile_df[(tile_df['datetime']==date_time) & (tile_df['asset']==band)]['file_path'].values[0])

        ### Flattening the file to get a vector for the image 
        array    = np.expand_dims(source.read(1).flatten(), axis=1)

        ### Capturing the date at which we need to replace the vector
        val      = date_dict[date_time]

        ### index at which replacement is to be done
        indices  = np.where(date_order_to_consider==val)[0][0]            

        X_tile[:,indices]   = array.ravel()

    return X_tile

def get_dataframe(data_dict_band,band,y,field_ids):
    
    X = np.array([values.tolist() for _,values in tqdm(data_dict_band.items())]).reshape(-1,76)
    colnames          = [band+'_'+z for z in date_order_to_consider]
    data              = pd.DataFrame(X,columns=colnames)
   
    data['field_id']  = field_ids
    
    mean_df           = data.groupby('field_id').mean().reset_index()
    low_df            = data.groupby('field_id').quantile(0.25).reset_index()
    up_df             = data.groupby('field_id').quantile(0.75).reset_index()
    med_df            = data.groupby('field_id').median().reset_index()
    
    
    return mean_df,low_df,up_df,med_df

len(tile_ids_train)

import xarray as xr
from rasterio.warp import transform

def convert_lat_lon(filename,crs):
    da = xr.open_rasterio(filename)
    x  = da['x']
    y  = da['y']
    ny, nx  = len(da['y']), len(da['x'])
    y, x    = np.meshgrid(da['y'], da['x'])
    
    lon,lat = transform(crs, {'init': 'EPSG:4326'},
                     x.flatten(), y.flatten())
    lon = np.asarray(lon).reshape((ny, nx))
    lat = np.asarray(lat).reshape((ny, nx))
    
    return lon,lat

lb = 0 
ub = 200

bands_available = ['B01']

bigdf = []

for band in bands_available:

    for batch in tqdm(range(0,6)):
        print(f'Performing operations for batch {batch+1}/14 for band {band}')
        data_dict_band = {}
        

        count          = 1

        y              = np.empty((0, 1))


        field_ids      = np.empty((0, 1))
        encode_x_val   = np.empty((0, 1))
        encode_y_val   = np.empty((0, 1))

        for tile_id in tile_ids_train[lb+(batch*200):ub+(batch*200)]:
            if tile_id != '1951': # avoid using this specific tile for the Hackathon as it might have a missing file

                tile_df               = competition_train_df[competition_train_df['tile_id']==tile_id]

                field_id_src          = rasterio.open(tile_df[tile_df['asset']=='field_ids']['file_path'].values[0])

                field_id_array        = field_id_src.read(1)
                
                lon,lat               = convert_lat_lon(tile_df[tile_df['asset']=='field_ids']['file_path'].values[0],field_id_src.crs)
                
                
                encode_x = np.zeros((256,256))
                encode_y = np.zeros((256,256))
                
                
                for i in range(field_id_array.shape[0]):
                    for j in range(field_id_array.shape[1]):
                        if field_id_array[i,j]>0:
                            encode_x[i,j] = lat[i][j]
                            encode_y[i,j] = lon[i][j]
                            
                tempdf = pd.DataFrame(columns = ['field_id','long','lat'])            
                tempdf['field_id'] = field_id_array.flatten()
                tempdf['long'] = encode_y.flatten()
                tempdf['lat'] = encode_x.flatten()
                tempdf      = tempdf[tempdf['field_id']!=0]
#                 print(tempdf.groupby(['field_id']).agg({'X':'median','Y':'median'}).reset_index().rename({'index':'field_id'}))
                gdf = tempdf.groupby(['field_id']).agg({'long':'median','lat':'median'}).reset_index().rename({'index':'field_id'})
                if len(bigdf)==0:
                    bigdf = gdf
                else:
                    bigdf = bigdf.append(gdf)
                    
                print(bigdf.shape)
                

                

                count                 = count+1
        gc.collect()        

#         mean_df,low_df,up_df,med_df = get_dataframe(data_dict_band,band,y,field_ids)
        
        batchid           = int(batch)+1

#         mean_df.to_csv(f'train_position.csv',index=False)
        
        

#         del field_ids,data_dict_band,label_src,field_id_src,label_array,field_id_array,y,mean_df,med_df,up_df,low_df
#         del tile_df,tile_date_times
#         gc.collect()
        gc.collect()
        gc.collect()

test_coordinates = bigdf[bigdf['field_id']!=0]

test_coordinates.to_csv('test_coordinates_lat_lon.csv',index=False)

test_coordinates
