==> 6th Place - Click Click Boom/zindi_sentinel2_codes/step7_nearest_points.ipynb <==
!pip install scikit-learn==1.0
!pip install xgboost==1.4.2
!pip install catboost==0.26.1
!pip install pandas==1.3.3
!pip install radiant-mlhub==0.3.0
!pip install rasterio==1.2.8
!pip install numpy==1.21.2
!pip install pathlib==1.0.1
!pip install tqdm==4.62.3
!pip install joblib==1.0.1
!pip install matplotlib==3.4.3
!pip install Pillow==8.3.2
!pip install torch==1.9.1
!pip install plotly==5.3.1

import pandas as pd
import numpy as np
from sklearn.neighbors import NearestNeighbors
from joblib import Parallel,delayed
import gc

train_df_mean = pd.read_csv('train_mean.csv')
train_df_coordinates = pd.read_csv('train_coordinates_lat_lon.csv')
test_df_coordinates  = pd.read_csv('test_coordinates_lat_lon.csv')

train_df_coordinates = train_df_coordinates.merge(train_df_mean,on=['field_id'],how='left')[['field_id','lat','long','label']]

train_df_coordinates = train_df_coordinates[train_df_coordinates['label'].isin(range(1,10))]


merge_df = train_df_coordinates.append(test_df_coordinates)
merge_df = merge_df.fillna(10)
merge_df.shape

field_label_dict = dict(zip(merge_df['field_id'].values,merge_df['label'].values))
field_range_dict = dict(zip(range(len(merge_df)),merge_df['field_id'].values))
len(field_label_dict)

### Nearest points within radius of 0.25

vals  = merge_df[['lat','long']].values
radius = 0.25

neigh       = NearestNeighbors(radius=radius,metric='haversine',n_jobs=-1).fit(vals)

def get_frequency(closest_indices):
    count_dict = {1:0,2:0,3:0,4:0,5:0,6:0,7:0,8:0,9:0,10:0}
    for i in closest_indices:

        label = field_label_dict[field_range_dict[i]]
        count_dict[label]+=1
    return [count_dict[i] for i in range(1,11)]

def get_nearest(point,radius):
    nbrs = neigh.radius_neighbors([point],radius=radius)
    indices = nbrs[1][0]
    return get_frequency(indices)




outs_train = Parallel(n_jobs=-1,timeout=100000,backend="multiprocessing", verbose=1)(delayed(get_nearest)(point=vals[i],radius=radius) for i in range(vals.shape[0]))

column_names = [f'Crop_{i}_{radius}' for i in range(1,11)]

nearest_data   = pd.DataFrame(data = outs_train,columns = column_names)
nearest_data['field_id'] = merge_df['field_id'].values.tolist()
nearest_data

nearest_data[f'count_{radius}'] = nearest_data[column_names].sum(axis=1)
for i in column_names:
    nearest_data[i] = 100*(nearest_data[i]/nearest_data[f'count_{radius}'])

nearest_data.to_csv(f'full_nearest_radius_{radius}.csv',index=False)

### Nearest data within radius of 0.4

vals  = merge_df[['lat','long']].values
radius = 0.4

neigh       = NearestNeighbors(radius=radius,metric='haversine',n_jobs=-1).fit(vals)

def get_frequency(closest_indices):
    count_dict = {1:0,2:0,3:0,4:0,5:0,6:0,7:0,8:0,9:0,10:0}
    for i in closest_indices:

        label = field_label_dict[field_range_dict[i]]
        count_dict[label]+=1
    return [count_dict[i] for i in range(1,11)]

def get_nearest(point,radius):
    nbrs = neigh.radius_neighbors([point],radius=radius)
    indices = nbrs[1][0]
    return get_frequency(indices)




outs_train = Parallel(n_jobs=-1,timeout=100000,backend="multiprocessing", verbose=1)(delayed(get_nearest)(point=vals[i],radius=radius) for i in range(vals.shape[0]))

column_names = [f'Crop_{i}_{radius}' for i in range(1,11)]

nearest_data   = pd.DataFrame(data = outs_train,columns = column_names)
nearest_data['field_id'] = merge_df['field_id'].values.tolist()
nearest_data

nearest_data[f'count_{radius}'] = nearest_data[column_names].sum(axis=1)
for i in column_names:
    nearest_data[i] = 100*(nearest_data[i]/nearest_data[f'count_{radius}'])

nearest_data

nearest_data.to_csv(f'full_nearest_radius_{radius}.csv',index=False)

# train_nearest = nearest_data[:train_df_coordinates.shape[0]]
# test_nearest  = nearest_data[train_df_coordinates.shape[0]:]

# # train_nearest.to_csv('nearest_data_train.csv',index=False)
# # test_nearest.to_csv('nearest_data_test.csv',index=False)
# train_nearest['label'] = train_df_coordinates['label'].values.tolist()
# train_nearest.shape,test_nearest.shape
