==> 6th Place - Click Click Boom/zindi_sentinel2_codes/step8_model1_akash.ipynb <==
!pip install scikit-learn==1.0
!pip install xgboost==1.4.2
!pip install catboost==0.26.1
!pip install pandas==1.3.3
!pip install radiant-mlhub==0.3.0
!pip install rasterio==1.2.8
!pip install numpy==1.21.2
!pip install pathlib==1.0.1
!pip install tqdm==4.62.3
!pip install joblib==1.0.1
!pip install matplotlib==3.4.3
!pip install Pillow==8.3.2
!pip install torch==1.9.1
!pip install plotly==5.3.1


gpu_info = !nvidia-smi
gpu_info = '\n'.join(gpu_info)
if gpu_info.find('failed') >= 0:
  print('Select the Runtime > "Change runtime type" menu to enable a GPU accelerator, ')
  print('and then re-execute this cell.')
else:
  print(gpu_info)

import pandas as pd
import numpy as np
import random
import torch
def seed_all(seed_value):
    random.seed(seed_value) # Python
    np.random.seed(seed_value) # cpu vars
    torch.manual_seed(seed_value) # cpu  vars
    
    if torch.cuda.is_available(): 
        torch.cuda.manual_seed(seed_value)
        torch.cuda.manual_seed_all(seed_value) # gpu vars
        torch.backends.cudnn.deterministic = True  #needed
        torch.backends.cudnn.benchmark = False

seed_all(13)

# from google.colab import drive
# drive.mount('/content/drive')

import warnings
warnings.filterwarnings("ignore")

import gc
import pandas as pd
import numpy as np
from sklearn.metrics import *
from xgboost import XGBClassifier
from catboost import CatBoostClassifier
from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import roc_auc_score
from sklearn.ensemble import VotingClassifier
from sklearn.linear_model import LogisticRegression
from indices_creation import *

## Data Load Step 
1. We load the mean aggregations for both train and test. The mean aggregations contain the labels and field IDs.
2. The quantile aggregations contain the field IDs. 

import os
os.getcwd()

train_df_mean = pd.read_csv('train_mean.csv')
#### we need to drop 'label' and 'field_id' later in the code 

test_df_mean  = pd.read_csv('test_mean.csv')
#### we need to drop 'field_id' later in the code 


train_df_median = pd.read_csv('train_median.csv')
#### we need to drop  'field_id' later in the code 

test_df_median  = pd.read_csv('test_median.csv')
#### we need to drop 'field_id' later in the code 

train_size    = pd.read_csv('size_of_field_train.csv')
test_size     = pd.read_csv('size_of_field_test.csv')

train_size    = train_size.rename({'Field_id':'field_id'},axis=1)
test_size     = test_size.rename({'Field_id':'field_id'},axis=1)

train_df_median = train_df_median.merge(train_size, on =['field_id'],how='left')
test_df_median  = test_df_median.merge(test_size, on =['field_id'],how='left')

cluster_df          = pd.read_csv('seven_cluster.csv')
cluster_df          = cluster_df.rename({'cluster_label':'cluster_label_7'},axis=1)

train_df_median = train_df_median.merge(cluster_df,on=['field_id'],how='left')
test_df_median  = test_df_median.merge(cluster_df,on=['field_id'],how='left')

gc.collect()

full_nearest1=pd.read_csv('full_nearest_radius_0.25.csv')
full_nearest2=pd.read_csv('full_nearest_radius_0.4.csv')

colsnearest40 = full_nearest2.columns.tolist()
_             = colsnearest40.remove('field_id')
colsnearest40

train_df_median      = train_df_median.merge(full_nearest1,on=['field_id'],how='left')
train_df_median      = train_df_median.merge(full_nearest2,on=['field_id'],how='left')
print(train_df_median.shape)
test_df_median       = test_df_median.merge(full_nearest1,on=['field_id'],how='left')
test_df_median       = test_df_median.merge(full_nearest2,on=['field_id'],how='left')


## Removing Erroneous data points
We observed some data points for which the labels were floats, we will remove them (they are few in number) to make sure our model is learning from correctly labelled data points

print(f'The shape of train data before outlier removal - {train_df_mean.shape}')

train_df_mean = train_df_mean[train_df_mean.label.isin(list(range(1,10)))]

print(f'The shape of train data after outlier removal - {train_df_mean.shape}')

relevant_fids   = train_df_mean['field_id'].values.tolist()
train_df_median = train_df_median[train_df_median['field_id'].isin(relevant_fids)]

print(f'The shape of median train data - {train_df_median.shape} and mean train data {train_df_mean.shape}' )

### two extra columns in train_df_mean being 'label' and 'size_of_field'

### Extract date list 
We extract the list of all dates where observations were seen for index generation

cols              = ['B01_','B02_','B03_','B04_','B05_','B06_','B07_','B08_','B09_','B8A_','B11_','B12_']

columns_available = train_df_mean.columns.tolist()

cols2consider     = []

for col in cols:
  cols2consider.extend( [c for c in columns_available if col in c])

bands_with_dates = [c for c in columns_available if 'B01_' in c]

dates            = [c.replace('B01_','') for c in bands_with_dates]

print(f'The sample showing the commencement dates where observations were seen is {dates[:10]}')
print(f'The sample showing the ending dates where observations were seen is {dates[-10:]}')

### Removal of field ID column
We consider only the relevant columns to be considered for the next step 

train_df_mean   = train_df_mean[cols2consider+['label']]
test_df_mean    = test_df_mean[cols2consider]

train_df_median = train_df_median[cols2consider+['size_of_field']+['cluster_label_7']+full_nearest1.columns.tolist()+colsnearest40]
test_df_median  = test_df_median[cols2consider+['size_of_field']+['cluster_label_7']+full_nearest1.columns.tolist()+colsnearest40]

### Indices Creation
We will create the indices for train and test data for mean aggregates using the indices coded in indices_creation.py module

# train_df_mean = get_band_ndvi_red(train_df_mean,dates)
# train_df_mean = get_band_afri(train_df_mean,dates)
# train_df_mean = get_band_evi2(train_df_mean,dates)
# train_df_mean = get_band_ndmi(train_df_mean,dates)
# train_df_mean = get_band_ndvi(train_df_mean,dates)
# train_df_mean = get_band_evi(train_df_mean,dates)
# train_df_mean = get_band_bndvi(train_df_mean,dates)
# train_df_mean = get_band_nli(train_df_mean,dates)
# train_df_mean = get_band_lci(train_df_mean,dates)


# test_df_mean = get_band_ndvi_red(test_df_mean,dates)
# test_df_mean = get_band_afri(test_df_mean,dates)
# test_df_mean = get_band_evi2(test_df_mean,dates)
# test_df_mean = get_band_ndmi(test_df_mean,dates)
# test_df_mean = get_band_ndvi(test_df_mean,dates)
# test_df_mean = get_band_evi(test_df_mean,dates)
# test_df_mean = get_band_bndvi(test_df_mean,dates)
# test_df_mean = get_band_nli(test_df_mean,dates)
# test_df_mean = get_band_lci(test_df_mean,dates)




We will create the indices for train and test data for median aggregates using the indices coded in indices_creation.py module

train_df_median = get_band_ndvi_red(train_df_median,dates)
train_df_median = get_band_afri(train_df_median,dates)
train_df_median = get_band_evi2(train_df_median,dates)
train_df_median = get_band_ndmi(train_df_median,dates)
train_df_median = get_band_ndvi(train_df_median,dates)
train_df_median = get_band_evi(train_df_median,dates)
train_df_median = get_band_bndvi(train_df_median,dates)
train_df_median = get_band_nli(train_df_median,dates)
# train_df_median = get_band_lci(train_df_median,dates)


test_df_median = get_band_ndvi_red(test_df_median,dates)
test_df_median = get_band_afri(test_df_median,dates)
test_df_median = get_band_evi2(test_df_median,dates)
test_df_median = get_band_ndmi(test_df_median,dates)
test_df_median = get_band_ndvi(test_df_median,dates)
test_df_median = get_band_evi(test_df_median,dates)
test_df_median = get_band_bndvi(test_df_median,dates)
test_df_median = get_band_nli(test_df_median,dates)
# test_df_median = get_band_lci(test_df_median,dates)


# train_df_median = train_df_median.drop(cols2consider,axis=1)
# test_df_median  = test_df_median.drop(cols2consider,axis=1)

train_df_mean.shape,train_df_median.shape,test_df_mean.shape,test_df_median.shape

######### Saving the label variable and dropping it from the data
train_y        = train_df_mean['label'].values
train_df_mean  = train_df_mean.drop(['label'],axis=1)


train_df_mean.replace([np.inf, -np.inf], np.nan, inplace=True)
test_df_mean.replace([np.inf, -np.inf], np.nan, inplace=True)

train_df_median.replace([np.inf, -np.inf], np.nan, inplace=True)
test_df_median.replace([np.inf, -np.inf], np.nan, inplace=True)

# train_df_slope.replace([np.inf, -np.inf], np.nan, inplace=True)
# test_df_slope.replace([np.inf, -np.inf], np.nan, inplace=True)

train = train_df_median.values
test  = test_df_median.values

# train         = pd.concat([train_df_median,train_df_slope],axis=1).values
# test          = pd.concat([test_df_median,test_df_slope],axis=1).values

print(f'The shape of model ready train data is {train.shape} and model ready test data is {test.shape}')
print(f'The shape of target is {train_y.shape}')

train1 = pd.read_csv('train_with_slopes.csv')
test1  = pd.read_csv('test_with_slopes.csv')

train1.replace([np.inf, -np.inf], np.nan, inplace=True)
test1.replace([np.inf, -np.inf], np.nan, inplace=True)



train2=pd.concat([pd.DataFrame(train1.values,columns=train1.columns),train_df_median[['size_of_field','cluster_label_7']+full_nearest1.columns.tolist()+colsnearest40].reset_index(drop=True)],axis=1)
test2=pd.concat([pd.DataFrame(test1.values,columns=test1.columns),test_df_median[['size_of_field','cluster_label_7']+full_nearest1.columns.tolist()+colsnearest40].reset_index(drop=True)],axis=1)
train2.head()

del train2['field_id']
del test2['field_id']

pivot=pd.read_csv('pivottable.csv')
pivot

train2=train2.merge(pivot,how='left',on='cluster_label_7')
test2=test2.merge(pivot,how='left',on='cluster_label_7')

train2

del train_df_mean,train_df_median,train1,train_size,test_df_mean,test_df_median,test1,test_size

import gc
gc.collect()

train = train2[list(set(train2.columns.tolist()))].values
test  = test2[list(set(train2.columns.tolist()))].values

train.shape,test.shape

# (1616-1520)/8

# train_slope = train1[[cols for cols in train1.columns if re.findall(r'\b(\w+NDVI_red_slope)\b',cols)]]
# test_slope  = test1[[cols for cols in test1.columns if re.findall(r'\b(\w+NDVI_red_slope)\b',cols)]]

# train         = pd.concat([train_df_median.reset_index(drop=True),train_slope],axis=1).values
# test          = pd.concat([test_df_median.reset_index(drop=True),test_slope],axis=1).values

# train.shape,test.shape


oof_pred               = np.zeros((len(train), 9))

y_pred_final           = np.zeros((len(test),9 ))

num_models             = 3
temperature            = 50

n_splits               = 15
error                  = []
kf                     = StratifiedKFold(n_splits=n_splits, shuffle=True, random_state=13)

for fold, (tr_ind, val_ind) in enumerate(kf.split(train, train_y)):
    wghts = [0]*num_models
    logloss = []

    X_train, X_val     = train[tr_ind], train[val_ind]
    # X_train1, X_val1   = train_max[tr_ind], train_max[val_ind]

    y_train, y_val     = train_y[tr_ind], train_y[val_ind]



    model1 = XGBClassifier(n_estimators=2000,random_state=13,learning_rate=0.04,colsample_bytree=0.95,reg_lambda=13,
                           
                           tree_method='gpu_hist',eval_metric='mlogloss')

    model2 = CatBoostClassifier(task_type='GPU',verbose=False,n_estimators=5000,random_state=13,auto_class_weights='SqrtBalanced',max_depth=9,learning_rate=0.05)
    
    model3 = CatBoostClassifier(task_type='GPU',verbose=False,n_estimators=5000,random_state=13,auto_class_weights='SqrtBalanced',max_depth=10,learning_rate=0.04)

#     model4 = CatBoostClassifier(task_type='GPU',verbose=False,n_estimators=5000,random_state=13,auto_class_weights='SqrtBalanced',max_depth=11)


    model1.fit(X_train,y_train)
    val_pred1 = model1.predict_proba(X_val)    
    logloss.append(log_loss(y_val,val_pred1))   
    print('validation logloss model1 fold-',fold+1,': ',log_loss(y_val,val_pred1))
   
    
    model2.fit(X_train,y_train)
    val_pred2 = model2.predict_proba(X_val) 
    logloss.append(log_loss(y_val,val_pred2))   
    print('validation logloss model2 fold-',fold+1,': ',log_loss(y_val,val_pred2))


    model3.fit(X_train,y_train)
    val_pred3 = model3.predict_proba(X_val) 
    logloss.append(log_loss(y_val,val_pred3))   
    print('validation logloss model3 fold-',fold+1,': ',log_loss(y_val,val_pred3))

#     model4.fit(X_train,y_train)
#     val_pred4 = model4.predict_proba(X_val) 
#     logloss.append(log_loss(y_val,val_pred4))   
#     print('validation logloss model4 fold-',fold+1,': ',log_loss(y_val,val_pred4))
    
    


    wghts = np.exp(-temperature*np.array(logloss/sum(logloss)))
    wghts = wghts/sum(wghts)
    print(wghts)

    val_pred           = wghts[0]*val_pred1+wghts[1]*val_pred2+wghts[2]*val_pred3 #+wghts[3]*val_pred4
    
    print('Validation logloss for fold- ',fold+1,': ',log_loss(y_val,val_pred))

    oof_pred[val_ind]  = val_pred
    
    y_pred_final += (wghts[0]*model1.predict_proba(test)+
                          wghts[1]*model2.predict_proba(test)+wghts[2]*model3.predict_proba(test)
                         )/(n_splits)

    
    

print('OOF LogLoss :- ',(log_loss(train_y,oof_pred)))

outputs               = y_pred_final.copy()

test_df               = pd.read_csv('test_mean.csv')
field_ids_test        = test_df['field_id'].values.tolist()

data_test             = pd.DataFrame(outputs)
data_test['field_id'] = field_ids_test
data_test             = data_test[data_test.field_id != 0]
data_test

data_test = data_test.rename(columns={
    0:'Lucerne/Medics',
    1:'Planted pastures (perennial)', 
    2:'Fallow',
    3:'Wine grapes',
    4:'Weeds',
    5:'Small grain grazing',
    6:'Wheat',
    7:'Canola',
    8:'Rooibos'
})

pred_df = data_test[['field_id', 'Lucerne/Medics', 'Planted pastures (perennial)', 'Fallow', 'Wine grapes', 'Weeds', 'Small grain grazing', 'Wheat', 'Canola', 'Rooibos']]
pred_df['field_id'] = pred_df['field_id'].astype(int)
pred_df = pred_df.sort_values(by=['field_id'],ascending=True)
pred_df

pred_df.to_csv('trial1_sep_akash.csv',index=False)
